Fixed pattern for server files
List of file patterns for server group was made for version 10 and above
and include pg_resetwal.exe, but not pg_resetxlog.exe, which caused version 9.6 installer to omit pg_resetxlog.exe program
PGPRO-2158.
Also it was noticed that unpackaged files left in the installation
directory
did not cause build failure, so this error went unnoticed for three month.